rename files and add tooltips
