+++ src/sentry/static/sentry/app/components/events/eventEntries.tsx
@@ -14,9 +14,9 @@ import EventDataSection from 'app/components/events/eventDataSection';
 import EventExtraData from 'app/components/events/eventExtraData/eventExtraData';
 import EventSdk from 'app/components/events/eventSdk';
 import EventTags from 'app/components/events/eventTags/eventTags';
+import EventVitals, {EventVitalContainer} from 'app/components/events/eventVitals';
 import EventGroupingInfo from 'app/components/events/groupingInfo';
 import EventPackageData from 'app/components/events/packageData';
-import EventVitals, {EventVitalContainer} from 'app/components/events/realUserMonitoring';
 import RRWebIntegration from 'app/components/events/rrwebIntegration';
 import EventSdkUpdates from 'app/components/events/sdkUpdates';
 import {DataSection} from 'app/components/events/styles';
