+++ src/sentry/static/sentry/app/components/events/eventVitals.tsx
